Former-commit-id: e4d362f74661d6cf3aaadb3221444e063efc5c10 [formerly c7da11e73d0facedcf7c9e45c20c1eabb30a9cb4]
Former-commit-id: c48174e159423cfa845a0685a6a397a85de799be

+++ deeplearning4j-core/src/main/java/org/deeplearning4j/util/OptimizerMatrix.java
@@ -1,6 +1,6 @@
 package org.deeplearning4j.util;
 
-import org.deeplearning4j.optimize.TrainingEvaluator;
+import org.deeplearning4j.optimize.api.TrainingEvaluator;
 
 /**
  * Optimizer that handles optimizing parameters. Handles line search
@@ -37,7 +37,7 @@ public interface OptimizerMatrix {
      * The tolerance for change when running
      * @param tolerance
      */
-    void setTolerance(double tolerance);
+    void setTolerance(float tolerance);
 
     /**
      * Sets the training evaluator
